Extract pending add store slice
